ozakharchuk: moved verifyProductPreUsedFilter() to FilterProductTest class
