+++ src/test/java/com/softserveinc/ita/rozetka/FilterProductTest.java
@@ -137,4 +137,47 @@ public class FilterProductTest extends TestRunner {
 
         softAssert.assertAll();
     }
+
+    @Test
+    public void verifyProductPreUsedFilter() {
+        SubcategoryPage subcategoryPage = homePage
+                .openCategoryPage(Category.LAPTOPS_AND_COMPUTERS)
+                .openSubcategoryPage(LaptopsAndComputersSubcategory.NOTEBOOKS);
+        Filter filter = subcategoryPage.getFilter();
+        filter.filter(ProductFilter.PRE_USED);
+        int productsQuantity = subcategoryPage.getProductsQuantity();
+        int productsQuantityToCheck = 30;
+
+        SoftAssertions softly = new SoftAssertions();
+
+        assertThat(productsQuantity)
+                .as("Products quantity should be sufficient")
+                .isGreaterThanOrEqualTo(productsQuantityToCheck);
+
+        for (int i = 1; i <= productsQuantity; i = i + 10) {
+            boolean isProductUsed = subcategoryPage
+                    .getProduct(i)
+                    .isUsed();
+            softly.assertThat(isProductUsed)
+                    .as("Product should be used")
+                    .isTrue();
+        }
+
+        subcategoryPage.resetFilters();
+        filter.filter(ProductFilter.NEW);
+
+        softly.assertThat(productsQuantity)
+                .as("Products quantity should be sufficient")
+                .isGreaterThanOrEqualTo(productsQuantityToCheck);
+
+        for (int i = 1; i <= productsQuantity; i = i + 10) {
+            boolean isProductUsed = subcategoryPage
+                    .getProduct(i)
+                    .isUsed();
+            softly.assertThat(isProductUsed)
+                    .as("Product should be new")
+                    .isFalse();
+        }
+        softly.assertAll();
+    }
 }
